☕ chore: update test suite
